+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changelog
 
 Breaking changes:
 
-- *add item here*
+- Move PTS Negotiator over form PlacelessTranslationService over to here in order to get rid of PTS itself.
+  [jensens, ksuess]
 
 New features: